% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_rank.R
\name{aft_rank}
\alias{aft_rank}
\title{Fit the DR-RML model}
\usage{
aft_rank(
  L,
  R,
  X,
  delta,
  id = NULL,
  alpha = 1,
  type = c("gehan", "logrank"),
  maxit = 20,
  tol = 1e-05,
  nboot = 0
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{X}{baseline covariate.}

\item{delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{alpha}{baseline covariate for fitting outcome regression model.}

\item{type}{type of rank estimation, \code{gehan}: gehan estimation; \code{logrank}: log-rank estimation.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 20.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-5.}

\item{nboot}{number of bootstrapped sample generation for variance estimation (Zeng and Lin, 2008) of regression estimator.}
}
\value{
\code{aft_rank} returns a data frame containing at least the following components:
\itemize{
  \item \code{est}: regression estimator.
  \item \code{se}: standard error estimates for \code{est}.
}
}
\description{
Estimate double-robust average causal effect for restricted mean lifetime.
}
\details{
see Choi et al., (2022+) for detailed method explanation.
}
\references{
Cao, W., Tsiatis, A. A., & Davidian, M. (2009). Improving efficiency and robustness of the doubly robust estimator for a population mean with incomplete data. \emph{Biometrika}, \bold{96}(3), 723--734.

Choi, S., Choi, T., Lee, H-Y., Han, S. W., Bandyopadhyay, D. (2022+). Double-robust methods for estimating differences in restricted mean lifetimes using pseudo-observations. In revision, \emph{Pharmaceutical Statistics}.
}
