% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_rank.R
\name{aft_rank}
\alias{aft_rank}
\title{rankIC}
\usage{
aft_rank(
  L,
  R,
  X,
  delta,
  id = NULL,
  alpha = 1,
  type = c("gehan", "logrank"),
  maxit = 20,
  tol = 1e-05,
  nboot = 0
)
}
\arguments{
\item{L}{left-censoring time, having 0 if left-censored.}

\item{R}{right-censoring time, having \code{Inf} if right-censored.}

\item{X}{baseline covariate.}

\item{delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{alpha}{baseline covariate for fitting outcome regression model.}

\item{type}{type of rank estimation, \code{gehan}: gehan estimation; \code{logrank}: log-rank estimation.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 20.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-5.}

\item{nboot}{number of bootstrapped sample generation for variance estimation of regression estimator (Zeng and Lin, 2008), 200 is usually recommended.}
}
\value{
\code{aft_rank} returns a data frame containing at least the following components:
\itemize{
  \item \code{est}: regression estimator.
  \item \code{se}: standard error estimates for \code{est}.
  \item \code{pvalue}: p-value, which will be appeared when the value of \code{nboot} is larger than one.
}
}
\description{
Fit the semiparametric accelerated failure time model with rank regression.
}
\details{
see Choi et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations


# Data example
library(PICBayes)
data(mCRC)
dt0 = as.data.frame(mCRC)
dt = with(dt0,
          data.frame(L = ifelse(is.na(L), 0, L),
                     R = ifelse(is.na(R), Inf, R),
                     delta = 1-IC,
                     x1 = TRT_C,
                     x2 = KRAS_C,
                     id = SITE))
d = dt
L = d$L; R = d$R; X = cbind(d$x1,d$x2); delta = d$delta; id = d$id
aft_rank(L,R,X,delta,id, alpha = 1, type = "gehan", nboot = 10)
aft_rank(L,R,X,delta,id, alpha = 1, type = "logrank", nboot = 10)
}
}
\references{
Zeng, D. and Lin, D. (2008). Efficient resampling methods for nonsmooth estimat- ing functions, Biostatistics 9(2): 355–363.

Choi, T., Choi, S. and Bandyopadhyay, D. (2022+). Rank estimation for the accelerated failure time model with partially interval-censored data. Submitted, \emph{Biometrics}.
}
