% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aft_rank.R
\name{aft_rank}
\alias{aft_rank}
\title{rankIC}
\usage{
aft_rank(
  U,
  V,
  X,
  Delta,
  id = NULL,
  alpha = 1,
  type = c("gehan", "logrank"),
  maxit = 20,
  tol = 1e-05,
  nboot = 0
)
}
\arguments{
\item{U}{left-censoring time, having 0 if left-censored.}

\item{V}{right-censoring time, having \code{Inf} if right-censored.}

\item{X}{baseline covariate.}

\item{Delta}{censoring indicator, 1: observed; 0: interval-censored.}

\item{id}{cluster id. If the data does not have clustered structure, set \code{id=NULL}.}

\item{alpha}{scalar to adjust informative cluster sizes, 0<alpha<=1. Cluster size will be ignored by using \code{alpha=0}.}

\item{type}{type of rank estimation, \code{gehan}: gehan estimation; \code{logrank}: log-rank estimation.}

\item{maxit}{maximum number of iteration for the log-rank estimator, default is 20.}

\item{tol}{tolerance of iteration for the log-rank estimator, default is 1e-5.}

\item{nboot}{number of bootstrapped sample generation for variance estimation of regression estimator (Zeng and Uin, 2008). For stable estimation, \code{nboot=200} is usually recommended.}
}
\value{
\code{aft_rank} returns a data frame containing at least the following components:
\itemize{
  \item \code{est}: regression estimator.
  \item \code{se}: standard error estimates for \code{est}.
  \item \code{pvalue}: p-value, which will be appeared when the value of \code{nboot} is larger than one.
}
}
\description{
Fit the semiparametric accelerated failure time model with rank regression.
}
\details{
see Choi et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
# Simulations
set.seed(111)
n = 200
x1 = rnorm(n)
x2 = ifelse(rnorm(n)>0, 1, 0)
X = cbind(x1,x2)
T = 2 + x1 + x2 + rnorm(n)
U = (1 - 0.25*x1)*runif(n, -6, 3.85)
V = U + (1 - 0.5*x2)*runif(n, 6, 19.6)
U = exp(dplyr::case_when(TRUE ~ T, T>V ~ V, T<U ~ -Inf))
V = exp(dplyr::case_when(TRUE ~ T, T>V ~ Inf, T<U ~ U))
Delta = ifelse(U==V, 1, 0)
aft_rank(U = U, V = V, X = X, Delta = Delta,
         type = "gehan", nboot = 10)
aft_rank(U = U, V = V, X = X, Delta = Delta,
         type = "logrank", nboot = 10)

# Data example
library(PICBayes)
data(mCRC)
dt0 = as.data.frame(mCRC)
d = with(dt0,
         data.frame(U = ifelse(is.na(L), 0, L),
                    V = ifelse(is.na(R), Inf, R),
                    Delta = 1-IC,
                    x1 = TRT_C,
                    x2 = KRAS_C,
                    id = SITE))
U = d$U; V = d$V; X = cbind(d$x1,d$x2); Delta = d$Delta; id = d$id
aft_rank(U = U, V = V, X = X, Delta = Delta, id = id, 
         alpha = 1, type = "gehan", nboot = 10)
aft_rank(U = U, V = V, X = X, Delta = Delta, id = id, 
         alpha = 1, type = "logrank", nboot = 10)
}
}
\references{
Zeng, D. and Lin, D. (2008). Efficient resampling methods for nonsmooth estimating functions. Biostatistics 9(2): 355–-363.

Choi, T., Choi, S. and Bandyopadhyay, D. (2022+). Rank estimation for the accelerated failure time model with partially interval-censored data. Submitted to Biometrics.

Pan, C. (2021). PICBayes: Bayesian Models for Partly Interval-Censored Data. R package. https://CRAN.R-project.org/package=PICBayes.
}
